refactor: split podcast creator into modular components
